🐛 Fix erroneous type annotations for `repeatList`
There's a potential case that the `append` function that it uses returns
a `Parser<null>` but we know that this usecase won't happen in our
scenario, because we feed it a sequence of copies of the original
parser, which is a `Parser<T>`. Thus we also know the return type of
`append` will always be a `Parser<T>`.

So we had to let Psalm know that we know this, and that the inference
that's being made is too eager.